requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0.0
